+++ docs/site/content/en/docs/overview/changelog.md
@@ -14,6 +14,8 @@ toc: true
 ---
 ## 2023.8
 August 2023
+- Added [status endpoints](https://waha.devlike.pro/docs/how-to/send-messages/#send-status-aka-stories) to **NOWEB** engine!
+  - `POST /api/{session}/status/text|image|voice|video` - choose one that suites your!
 - Added [chats endpoints](https://waha.devlike.pro/docs/how-to/chats)
   - Get all chats - `GET /api/{session/chats`
   - Delete chat - `DELETE /api/{session/chats/{chatId}`